requirements.txt
scikit-learn
scikit-optimize
matplotlib
python-box==6.0.2
pyYAML
tqdm
pydantic
ensure==1.0.2
joblib
types-PyYAML
fastapi
uvicorn
mlflow
xgboost
seaborn
streamlit
dagshub
docker
tensorflow
prometheus-client
dvc
statsmodels
numpy
pandas
imblearn
-e .